[Enhancement #520] Move shrinkFullIri to Utils
Moving shrinkFullIri to Utils allows consistent IRI shrinking outside the AssetLabel component.
